scrapper/src/main/java/edu/java/dto/GitHubRepositoryDTO.java
package edu.java.dto;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.json.JsonMapper;
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
import java.io.IOException;
import java.time.OffsetDateTime;

@JsonIgnoreProperties(ignoreUnknown = true)
public class GitHubRepositoryDTO {
    private long id;
    @JsonProperty("full_name")
    private String fullName;
    @JsonProperty("updated_at")
    private OffsetDateTime updatedAt;

    public static GitHubRepositoryDTO fromJson(String json) throws IOException {
        ObjectMapper mapper = JsonMapper.builder().addModule(new JavaTimeModule()).build();
        return mapper.readValue(json, GitHubRepositoryDTO.class);
    }

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getFullName() {
        return fullName;
    }

    public void setFullName(String fullName) {
        this.fullName = fullName;
    }

    public OffsetDateTime getUpdatedAt() {
        return updatedAt;
    }

    public void setUpdatedAt(OffsetDateTime updatedAt) {
        this.updatedAt = updatedAt;
    }
}
